+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## UNRELEASED - tbc
+## [0.6.0] - 2026-09-10
 
 ### Added
 
@@ -98,6 +98,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 Initial release
 
 
+[0.6.0]: https://github.com/elliot-100/british-cycling-utils/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/elliot-100/british-cycling-utils/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/elliot-100/british-cycling-utils/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/elliot-100/british-cycling-utils/compare/v0.3.1...v0.4.0